requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.1.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.1.0=pypi_0
blas=1.0=mkl
blessings=1.7=pypi_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
charset-normalizer=2.0.12=pypi_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pypi_0
decorator=4.4.2=pypi_0
faiss-gpu=1.7.1=py3.6_h293177f_1_cuda11.0
frozenlist=1.2.0=pypi_0
fsspec=2022.1.0=pypi_0
future=0.18.2=pypi_0
google-auth=2.8.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gpustat=0.6.0=pypi_0
gputil=1.4.0=pypi_0
grpcio=1.46.3=pypi_0
idna=3.3=pypi_0
idna-ssl=1.1.0=pypi_0
imageio=2.15.0=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2022.1.0=h9e868ea_3769
joblib=1.1.0=pypi_0
kiwisolver=1.3.1=pypi_0
libedit=3.1.20210910=h7f8727e_0
libfaiss=1.7.1=h7f34bec_1_cuda11.0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.36.0=pypi_0
markdown=3.3.7=pypi_0
matplotlib=3.3.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
mmcv-full=1.5.3=pypi_0
multidict=5.2.0=pypi_0
ncurses=6.3=h7f8727e_2
networkx=2.5.1=pypi_0
numba=0.53.1=pypi_0
numpy=1.19.5=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.2.0=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.1.5=pypi_0
pillow=8.4.0=pypi_0
pip=21.2.2=py36h06a4308_0
pluggy=1.0.0=pypi_0
progress=1.6=pypi_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
py=1.11.0=pypi_0
pyamg=4.2.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeprecate=0.3.1=pypi_0
pynndescent=0.5.10=pypi_0
pyparsing=3.0.7=pypi_0
pytest=7.0.1=pypi_0
python=3.6.7=h0371630_0
python-dateutil=2.8.2=pypi_0
pytorch-warmup=0.0.4=pypi_0
pytz=2022.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=6.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.5.4=pypi_0
seaborn=0.11.2=pypi_0
setuptools=59.5.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.5.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2020.9.3=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=1.2.3=pypi_0
torch=1.10.1+cu111=pypi_0
torchaudio=0.10.1+rocm4.1=pypi_0
torchmetrics=0.8.2=pypi_0
torchvision=0.11.2+cu111=pypi_0
tqdm=4.64.0=pypi_0
typing-extensions=4.1.1=pypi_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.9=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yapf=0.32.0=pypi_0
yarl=1.7.2=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.12=h7f8727e_2
